chore: convert some custom ZRC model classes to Kotlin
