+++ src/main/kotlin/nl/info/client/zgw/zrc/model/ZaakUuid.kt
@@ -2,7 +2,7 @@
  * SPDX-FileCopyrightText: 2024 INFO.nl
  * SPDX-License-Identifier: EUPL-1.2+
  */
-package net.atos.client.zgw.zrc.model
+package nl.info.client.zgw.zrc.model
 
 import nl.info.zac.util.AllOpen
 import nl.info.zac.util.NoArgConstructor
